refactor project migration
